requirements.txt
streamlit>=1.28.0
google-generativeai>=0.3.0
Pillow>=10.0.0
python-dotenv>=1.0.0
langsmith>=0.1.0
langchain>=0.1.0
langgraph>=0.2.0
langchain-google-genai>=1.0.0
requests>=2.31.0
httpx>=0.25.0
beautifulsoup4>=4.12.0
